+++ R/zzz.R
@@ -67,6 +67,10 @@
     "stability",
     "y1",
     "y2",
-    "n_mean"
+    "n_mean",
+    "Lonm",
+    "Latm",
+    "Am", 
+    "Bm"
   ))
 }